refactor(GiniBankSDK): Refactor setup properties for integration tests to be reused in other tests and remove duplicated code
PP-338
