+++ BankSDK/GiniBankSDKExample/Tests/DigitalInvoiceViewControllerTests.swift
@@ -10,23 +10,12 @@ import XCTest
 @testable import GiniBankSDK
 
 class DigitalInvoiceIntegrationTests: XCTestCase {
-    let clientId = ProcessInfo.processInfo.environment["CLIENT_ID"]!
-    let clientSecret = ProcessInfo.processInfo.environment["CLIENT_SECRET"]!
-    var giniBankAPILib: GiniBankAPI!
-    var giniCaptureSDKDocumentService: GiniCaptureSDK.DocumentService!
-    var giniBankAPIDocumentService: GiniBankAPILibrary.DefaultDocumentService!
+
+    lazy var giniHelper = GiniSetupHelper()
 
     override func setUp() {
-        let client = Client(id: clientId,
-                            secret: clientSecret,
-                            domain: "bank-sdk-example")
-        giniBankAPILib = GiniBankAPI
-            .Builder(client: client)
-            .build()
 
-        giniCaptureSDKDocumentService = DocumentService(lib: giniBankAPILib, metadata: nil)
-
-        giniBankAPIDocumentService = giniBankAPILib.documentService()
+        giniHelper.setup()
     }
 
     func testExtractionsForLineItemsWithDiscount() {
@@ -54,6 +43,7 @@ class DigitalInvoiceIntegrationTests: XCTestCase {
                 XCTFail("Error loading file: `\(fileName).json`")
                 return
             }
+
             let fixtureExtractionsContainer = try? JSONDecoder().decode(ExtractionsContainer.self, from: fixtureExtractionsJson)
 
             // 2. Verify we received the correct extractions for this test
@@ -130,14 +120,14 @@ class DigitalInvoiceIntegrationTests: XCTestCase {
         let captureDocument = builder.build(with: testDocumentData,
                                             fileName: "\(fileName).pdf")!
 
-        GiniBankConfiguration.shared.documentService = giniCaptureSDKDocumentService
+        GiniBankConfiguration.shared.documentService = giniHelper.giniCaptureSDKDocumentService
 
         // Upload a test document
-        giniCaptureSDKDocumentService.upload(document: captureDocument) { result in
+        giniHelper.giniCaptureSDKDocumentService.upload(document: captureDocument) { result in
             switch result {
                 case .success(_):
                     // Analyze the uploaded test document
-                    self.giniCaptureSDKDocumentService?.startAnalysis { result in
+                    self.giniHelper.giniCaptureSDKDocumentService?.startAnalysis { result in
                         switch result {
                             case let .success(extractionResult):
                                 let extractions: [String: Extraction] = Dictionary(uniqueKeysWithValues: extractionResult.extractions.compactMap {
@@ -149,7 +139,7 @@ class DigitalInvoiceIntegrationTests: XCTestCase {
                                 let analysisResult = AnalysisResult(extractions: extractions,
                                                                     lineItems: extractionResult.lineItems,
                                                                     images: [],
-                                                                    document: self.giniCaptureSDKDocumentService?.document,
+                                                                    document: self.giniHelper.giniCaptureSDKDocumentService?.document,
                                                                     candidates: [:])
 
                                 delegate.giniCaptureAnalysisDidFinishWith(result: analysisResult)
